Fix some C++ compiler warnings to comply with -Wall.
Some progress towards #311.

--
Reviewed-on: https://bazel-review.googlesource.com/#/c/1710
MOS_MIGRATED_REVID=99019961

+++ src/main/native/unix_jni.cc
@@ -632,7 +632,7 @@ Java_com_google_devtools_build_lib_unix_FilesystemUtils_readdir(JNIEnv *env,
     return NULL;  // async exception!
   }
 
-  for (int ii = 0; ii < len; ++ii) {
+  for (size_t ii = 0; ii < len; ++ii) {
     jstring s = NewStringLatin1(env, entries[ii].c_str());
     if (s == NULL && env->ExceptionOccurred()) {
       return NULL;  // async exception!
